backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
pydantic-settings>=2.5.2
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
httpx>=0.27.2
slowapi>=0.1.9
email-validator>=2.0.0
aiosqlite>=0.20.0
yookassa>=3.0.0
